refactor(argus.backend): Refactor cli commands, remove unused methods
